requirements.txt
flask
gunicorn
Jinja2
Keras
pandas
Pillow
tensorflow
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib